build: fix Makefile variables for MSYS2 and Cygwin
They were not taken into account before. This resulted in (seemingly harmless
but annoying) complaints about winepath being missing, but the build would
eventually fail when invoking makelcf, as the path to the template file was
invalid (as it would try to expand the PROJECT_ROOT_NT variable, which
was incorrectly blank, ultimately due to the wrong WINPATH).

Fixes #4